Remove GetStudentLessons from LessonsController
