+++ CharlieBackend.Api/Controllers/LessonsController.cs
@@ -88,22 +88,6 @@ namespace CharlieBackend.Api.Controllers
         }
 
         /// <summary>
-        /// Returns list of lessons of exact student
-        /// </summary>
-        /// <param name="id"></param>
-        /// <response code="200">Successful return of lessons list of given student</response>
-        [SwaggerResponse(200, type: typeof(IList<StudentLessonDto>))]
-        [Authorize(Roles = "Admin, Mentor, Secretary, Student")]
-        [HttpGet("students/{id}")]
-        public async Task<ActionResult<List<StudentLessonDto>>> GetStudentLessons(long id)
-        {
-            var lessons = await _lessonService.GetStudentLessonsAsync(id);
-
-            return lessons.ToActionResult();
-        }
-
-        
-        /// <summary>
         /// Updates given lesson
         /// </summary>
         /// <remarks>
